requirements.txt
zyte-api>=0.3.0
requests>=2.28.0
python-dotenv>=0.19.0
beautifulsoup4>=4.9.3
parsel>=1.6.0
python-json-logger>=2.0.0 